Inline enable-embedding section under its matching radio
